requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
pydantic==2.12.5
pydantic-settings==2.12.0
onnxruntime==1.23.2
numpy==2.2.6
pillow==12.0.0
opencv-python-headless==4.12.0.88
pandas==2.3.3
pytest==9.0.1
python-dotenv==1.2.1
